refactor: conceded to use a subdomain instead of a subfolder for the project.
I wanted silentmode.tv/real-figures, but it was a headache to try and set up on the server, and would be a headache to maintain in the future (as the main site is a symlink).
The subdomain method - realfigures.silentmode.tv - will allow the project to be updated independently of the main site.